persona por id

+++ src/main/java/com/itgt/pos/manager/PersonaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.itgt.pos.model.Persona;
 
-public interface PersonaRepository extends JpaRepository<Persona, Integer>{
+public interface PersonaRepository extends JpaRepository<Persona, Long>{
 
 }
